--- apps/asana/src/inngest/middlewares/rate-limit-middleware.ts
@@ -1,38 +1,0 @@
-import { InngestMiddleware, RetryAfterError } from 'inngest';
-import { AsanaError } from '@/connectors/common/error';
-
-export const rateLimitMiddleware = new InngestMiddleware({
-  name: 'rate-limit',
-  init: () => {
-    return {
-      onFunctionRun: ({ fn }) => {
-        return {
-          transformOutput: (ctx) => {
-            const {
-              result: { error, ...result },
-              ...context
-            } = ctx;
-
-            if (error instanceof AsanaError && error.response?.status === 429) {
-              const retryAfter = error.response.headers.get('retry-after') || 60;
-
-              return {
-                ...context,
-                result: {
-                  ...result,
-                  error: new RetryAfterError(
-                    `Rate limit exceeded for '${fn.name}'. Retry after ${retryAfter} seconds.`,
-                    `${retryAfter}s`,
-                    {
-                      cause: error,
-                    }
-                  ),
-                },
-              };
-            }
-          },
-        };
-      },
-    };
-  },
-});
